// generated/graphql/types.d.ts
export type Maybe<T> = T | undefined;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Query = {
  __typename?: 'Query';
  houses: Array<House>;
  people?: Maybe<Array<Person>>;
};

export type House = {
  __typename?: 'House';
  id: Scalars['Int'];
  squarefeet: Scalars['Int'];
  numberOfBedrooms: Scalars['Int'];
  onSale: Scalars['Boolean'];
  isRental?: Maybe<Scalars['Boolean']>;
  Owner: Person;
  address: HouseAddress;
};

export type HouseAddress = {
  __typename?: 'HouseAddress';
  streetname?: Maybe<Scalars['String']>;
  housenumber?: Maybe<Scalars['String']>;
};

export type Person = {
  __typename?: 'Person';
  id: Scalars['Int'];
  firstname: Scalars['String'];
  lastname: Scalars['String'];
  fullname: Scalars['String'];
  houses?: Maybe<Array<Maybe<House>>>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type LegacyStitchingResolver<TResult, TParent, TContext, TArgs> = {
  fragment: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type NewStitchingResolver<TResult, TParent, TContext, TArgs> = {
  selectionSet: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type StitchingResolver<TResult, TParent, TContext, TArgs> = LegacyStitchingResolver<TResult, TParent, TContext, TArgs> | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type IsTypeOfResolverFn<T = {}> = (obj: T, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Query: ResolverTypeWrapper<{}>;
  House: ResolverTypeWrapper<Partial<House>>;
  Int: ResolverTypeWrapper<Partial<Scalars['Int']>>;
  Boolean: ResolverTypeWrapper<Partial<Scalars['Boolean']>>;
  HouseAddress: ResolverTypeWrapper<Partial<HouseAddress>>;
  String: ResolverTypeWrapper<Partial<Scalars['String']>>;
  Person: ResolverTypeWrapper<Partial<Person>>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Query: {};
  House: Partial<House>;
  Int: Partial<Scalars['Int']>;
  Boolean: Partial<Scalars['Boolean']>;
  HouseAddress: Partial<HouseAddress>;
  String: Partial<Scalars['String']>;
  Person: Partial<Person>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  houses?: Resolver<Array<ResolversTypes['House']>, ParentType, ContextType>;
  people?: Resolver<Maybe<Array<ResolversTypes['Person']>>, ParentType, ContextType>;
};

export type HouseResolvers<ContextType = any, ParentType extends ResolversParentTypes['House'] = ResolversParentTypes['House']> = {
  id?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  squarefeet?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  numberOfBedrooms?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  onSale?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
  isRental?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  Owner?: Resolver<ResolversTypes['Person'], ParentType, ContextType>;
  address?: Resolver<ResolversTypes['HouseAddress'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType>;
};