// generated/grpcServer/person_pb.d.ts
    export type AsObject = {
        id: number,
    }
// ...
export class Person extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setId(value?: google_protobuf_wrappers_pb.Int32Value): Person;


    hasFirstname(): boolean;
    clearFirstname(): void;
    getFirstname(): google_protobuf_wrappers_pb.StringValue | undefined;
    setFirstname(value?: google_protobuf_wrappers_pb.StringValue): Person;


    hasLastname(): boolean;
    clearLastname(): void;
    getLastname(): google_protobuf_wrappers_pb.StringValue | undefined;
    setLastname(value?: google_protobuf_wrappers_pb.StringValue): Person;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Person.AsObject;
    static toObject(includeInstance: boolean, msg: Person): Person.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: Person, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Person;
    static deserializeBinaryFromReader(message: Person, reader: jspb.BinaryReader): Person;
}

export namespace Person {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        firstname?: google_protobuf_wrappers_pb.StringValue.AsObject,
        lastname?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}